% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_terms.R
\name{abbreviate_terms}
\alias{abbreviate_terms}
\title{abbreviate labels}
\usage{
abbreviate_terms(x)
}
\arguments{
\item{x}{vector of labels}
}
\value{
vector of abbreviations
}
\description{
Takes first letter of each word, where a word is separated by space on hyphen
Also removes brackets first
}
\examples{
abbreviate_terms(c("Jono Tuke", "Adam Ben Rohrlach", "Adelaide-University"))
}
