% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alpha.R
\name{add_alpha}
\alias{add_alpha}
\title{Add alpha}
\usage{
add_alpha(g, measure = NULL)
}
\arguments{
\item{g}{igraph network}

\item{measure}{centrality measure to set .alpha to}
}
\value{
igraph network with an vertex attribute called .alpha set to give centrality
measure that can be used in \link{plot_ggnet}
}
\description{
Add alpha
}
\examples{
add_alpha(example_network, "degree")
}
