% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_convert_bb_adj.R
\name{add_convert_bb_adj}
\alias{add_convert_bb_adj}
\title{Adjust BB for network work sf object}
\usage{
add_convert_bb_adj(network_sf, asp = 1)
}
\arguments{
\item{network_sf}{network sf objects}

\item{asp}{aspect ratio}
}
\value{
same network sf with adjusted aspect ration
}
\description{
Adjust BB for network work sf object
}
\examples{
add_convert_bb_adj(example_sf)
}
