% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ibd_file.R
\name{check_ibd_file}
\alias{check_ibd_file}
\title{Check IBD file}
\usage{
check_ibd_file(file)
}
\arguments{
\item{file}{IBD TSV file}
}
\value{
Message about whether TSV has correct columns
}
\description{
Check IBD file
}
