% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ibd_file.R
\name{check_ibd_files}
\alias{check_ibd_files}
\title{check IBD files}
\usage{
check_ibd_files(ibd, meta)
}
\arguments{
\item{ibd}{ibd file}

\item{meta}{meta file}
}
\value{
messages about if the files are fine
}
\description{
check IBD files
}
