% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ibd_file.R
\name{check_meta_file}
\alias{check_meta_file}
\title{Check meta file}
\usage{
check_meta_file(file)
}
\arguments{
\item{file}{IBD TSV meta file}
}
\value{
Message about whether TSV has correct columns
}
\description{
Check meta file
}
