% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_models.R
\name{clean_models}
\alias{clean_models}
\title{clean model names}
\usage{
clean_models(x)
}
\arguments{
\item{x}{vector of ergm model names}
}
\value{
cleaned vector
}
\description{
clean model names
}
\examples{
ergms <- example_network |>
  get_ergms(c("site", "genetic_sex")) |>
  purrr::map(broom::glance) |>
  purrr::list_rbind(names_to = "model")
clean_models(ergms$model) |> print()
}
