% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_network.R
\name{clean_network}
\alias{clean_network}
\title{Clean network}
\usage{
clean_network(g, col)
}
\arguments{
\item{g}{IBD igraph obj}

\item{col}{node attribute}
}
\value{
cleaned IBD igraph obj
}
\description{
Removes all nodes with NAs for a given attribute
so can fit ERGM models
}
\examples{
clean_network(example_network_3, "site")
}
