% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_convert_bb_adj.R
\name{convert_bb}
\alias{convert_bb}
\title{Given a bounding box, increase the lat and long by given amounts}
\usage{
convert_bb(bb, x_adj = 0.1, y_adj = 0.1)
}
\arguments{
\item{bb}{bounding box (left, bottom, right, top)}

\item{x_adj}{lon adjust}

\item{y_adj}{lat adjust}
}
\value{
new bounding box
}
\description{
Given a bounding box, increase the lat and long by given amounts
}
