% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_convert_bb_adj.R
\name{convert_bb_asp}
\alias{convert_bb_asp}
\title{convert bb to be square}
\usage{
convert_bb_asp(bb, asp = 1)
}
\arguments{
\item{bb}{bounding box}

\item{asp}{aspect ratio, default 1}
}
\value{
adjusted BB
}
\description{
convert bb to be square
}
