% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pipe.R
\name{convert_pipe}
\alias{convert_pipe}
\title{convert pipe}
\usage{
convert_pipe(x)
}
\arguments{
\item{x}{string for inclusion exclusion}
}
\value{
commas are replaced by pipe
}
\description{
convert pipe
}
\examples{
convert_pipe("bob, harry")
}
