% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ergms.R
\name{convert_pred_ergmterm}
\alias{convert_pred_ergmterm}
\title{convert_pred_terms}
\usage{
convert_pred_ergmterm(pred, g)
}
\arguments{
\item{pred}{name of igraph vertex attribute}

\item{g}{igraph object}
}
\value{
ergm term for attribute based on class
}
\description{
convert_pred_terms
}
\examples{
convert_pred_ergmterm("site", example_network)
}
