% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ringbauer_measures.R
\name{convert_ringbauer_measures}
\alias{convert_ringbauer_measures}
\title{Convert ringbauer measures}
\usage{
convert_ringbauer_measures(RM, abbr = TRUE)
}
\arguments{
\item{RM}{A ringbauer measure tibble}

\item{abbr}{a boolean that when true will shorten the group names}
}
\value{
a list of three matrices: density, labels, and text colour
}
\description{
This converts the ringbauer measures into matrices of the correct form
for the plot function.
}
\examples{
get_ringbauer_measures(example_network, "site") |>
convert_ringbauer_measures()
}
