% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sf.R
\name{convert_sf}
\alias{convert_sf}
\title{Convert to SF}
\usage{
convert_sf(g, lat, lon, col = "site")
}
\arguments{
\item{g}{igraph network}

\item{lat}{attribute that gives latitude}

\item{lon}{attribute that gives longitude}

\item{col}{attribute that will give colour in leaflet or plot}
}
\value{
list of edges_sf and nodes_sf
}
\description{
Converts an igraph object into two sf objects
}
\examples{
convert_sf(example_network,"lat","long",col = "site")
}
