% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{copy_files}
\alias{copy_files}
\title{Copy files}
\usage{
copy_files(folder)
}
\arguments{
\item{folder}{folder to place files}
}
\value{
copies files
}
\description{
Small internal function for me that puts example IBD
files and meta files into a folder to test
spIBDer_app
}
