% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-wizard.R
\name{dataWizardInput}
\alias{dataWizardInput}
\title{dataWizardInput}
\usage{
dataWizardInput(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
input UI for data Wizard
}
\description{
dataWizardInput
}
