% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-default-image.R
\docType{data}
\name{default_image}
\alias{default_image}
\title{default image}
\format{
An igraph object
\describe{
\item{row}{row}
\item{col}{column}
\item{pixel}{pixel}
}
}
\source{
1980s
}
\usage{
default_image
}
\description{
A single image as place holder
}
\keyword{datasets}
