% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_network.R
\docType{data}
\name{dist_network}
\alias{dist_network}
\title{dist network}
\format{
An igraph object
\describe{
\item{name}{Name for each node}
\dots
\item{degree}{Degree for each node}
\item{wij}{Weight for each edge}
}
}
\source{
simulated using erdos-renyi
}
\usage{
dist_network
}
\description{
An igraph example network that is used to test create_network function
}
\keyword{datasets}
