% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_to_sf.R
\name{edges_to_sf}
\alias{edges_to_sf}
\title{edges to sf}
\usage{
edges_to_sf(graph, lat = "lat", lon = "long")
}
\arguments{
\item{graph}{igraph obj}

\item{lat}{lat attribute}

\item{lon}{long attribute}
}
\value{
sf edge object
}
\description{
edges to sf
}
\examples{
edges_to_sf(example_network)
}
