% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_network.R
\docType{data}
\name{example_network}
\alias{example_network}
\title{example network}
\format{
An igraph object
\describe{
\item{site}{Site for each sample with levels A, B, or C}
\item{genetic}{Genetic sex for each sample with levels F or M}
}
}
\source{
Simulation via ergm package
}
\usage{
example_network
}
\description{
An igraph example network
}
\keyword{datasets}
