% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_network_2.R
\docType{data}
\name{example_network_2}
\alias{example_network_2}
\title{example network 2}
\format{
An igraph object
\describe{
\item{name}{Name for each node}
\dots
\item{degree}{Degree for each node}
\item{wij}{Weight for each edge}
}
}
\source{
Example IBD datasets from Dr Adam Rohrlach
}
\usage{
example_network_2
}
\description{
An igraph example network
}
\keyword{datasets}
