% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_network_3.R
\docType{data}
\name{example_network_3}
\alias{example_network_3}
\title{example network 2}
\format{
An igraph object
\describe{
\item{name}{Name for each node}
\dots
\item{degree}{Degree for each node}
\item{wij}{Weight for each edge}
}
}
\source{
Based on example_network
}
\usage{
example_network_3
}
\description{
An igraph example network
This one has missing node attributes
}
\keyword{datasets}
