% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sf.R
\docType{data}
\name{example_sf}
\alias{example_sf}
\title{example sf object}
\format{
Two SF objects
\describe{
\item{nodes_sf}{SF object for nodes}
\item{ndges_sf}{SF object for edges}
}
}
\source{
Simulation via ergm package
}
\usage{
example_sf
}
\description{
The example network as an sf object
}
\keyword{datasets}
