% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_convert_bb_adj.R
\name{expand_line}
\alias{expand_line}
\title{Expand a line by given amount}
\usage{
expand_line(x, k = 0.1)
}
\arguments{
\item{x}{2 vector with left and right point of a line}

\item{k}{amount to increase by}
}
\value{
2 vector of new left and right point
}
\description{
Expand a line by given amount
}
