% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_palettes.R
\name{fill_continuous}
\alias{fill_continuous}
\title{fill continuous}
\usage{
fill_continuous(pal = "ravenclaw")
}
\arguments{
\item{pal}{palette to use}
}
\value{
appropriate scale
}
\description{
fill continuous
}
