% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_palettes.R
\name{fill_discrete}
\alias{fill_discrete}
\title{fill discrete}
\usage{
fill_discrete(pal = "ravenclaw")
}
\arguments{
\item{pal}{palette to use}
}
\value{
appropriate scale
}
\description{
fill discrete
}
