% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sf.R
\name{filter_sf}
\alias{filter_sf}
\title{Filter SF object}
\usage{
filter_sf(sf_obj, xmin, xmax, ymin, ymax)
}
\arguments{
\item{sf_obj}{object with node and edge df}

\item{xmin}{left long border}

\item{xmax}{right long border}

\item{ymin}{bottom lat border}

\item{ymax}{top lat border}
}
\value{
filter SF object
}
\description{
Filter SF object
}
