% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ergms.R
\name{get_all_models}
\alias{get_all_models}
\title{get_all_models}
\usage{
get_all_models(preds, y, constant = "edges")
}
\arguments{
\item{preds}{vector of igraph attributes}

\item{y}{outcome variable}

\item{constant}{constant term}
}
\value{
a vector of strings of linear models
}
\description{
this function returns all possible models with the outcome variable y, and the predictors preds.
}
\examples{
get_all_models(c("X1", "X2"), "y", "1")
}
