% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-filter-network.R
\name{get_attr_values}
\alias{get_attr_values}
\title{Get the observed values of a node or edge attribute}
\usage{
get_attr_values(g, x, what = c("node", "edge"))
}
\arguments{
\item{g}{An igraph object.}

\item{x}{Name of the attribute.}

\item{what}{Either \code{"node"} or \code{"edge"}.}
}
\description{
Get the observed values of a node or edge attribute
}
