% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centrality_measures.R
\name{get_centrality_measures}
\alias{get_centrality_measures}
\title{Get centrality measures}
\usage{
get_centrality_measures(g, var = NULL)
}
\arguments{
\item{g}{igraph object}

\item{var}{character vector of attributes to stratify on}
}
\value{
tibble of centrality measures
}
\description{
Gives a table of mean centrality measures
given a list of vertex attributes to stratify
on. If not list given then gives overall.
}
\examples{
get_centrality_measures(example_network, c("site", "genetic_sex"))
}
