% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected_graph.R
\name{get_connected_graph}
\alias{get_connected_graph}
\title{get_connected_graph}
\usage{
get_connected_graph(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
igraph object
}
\description{
get_connected_graph
}
\examples{
g <- igraph::sample_gnp(10, 0.5)
get_connected_graph(g)
}
