% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_info.R
\name{get_edge_info}
\alias{get_edge_info}
\title{Get network edge information}
\usage{
get_edge_info(g)
}
\arguments{
\item{g}{ibd network object}
}
\value{
tibble of node information
}
\description{
Get network edge information
}
\examples{
get_edge_info(example_network_2)
}
