% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ergm.R
\name{get_ergm_bic}
\alias{get_ergm_bic}
\title{Tabulate the ergm AIC and BIC}
\usage{
get_ergm_bic(ergms)
}
\arguments{
\item{ergms}{list of ergm fits}
}
\value{
tibble of AIC and BIC for each model
}
\description{
Tabulate the ergm AIC and BIC
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex")
)
ergms |> get_ergm_bic()
}
