% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ergm_coef.R
\name{get_ergm_coef}
\alias{get_ergm_coef}
\title{Gets coefficients and fold changes for a ergm model}
\usage{
get_ergm_coef(ergm)
}
\arguments{
\item{ergm}{ergm model}
}
\value{
coefficient table
}
\description{
Gets coefficients and fold changes for a ergm model
}
\examples{
ergms <- get_ergms(
example_network,
preds = c("site", "genetic_sex"),
types = c("nodematch", "nodematch")
)
ergms[[1]] |> get_ergm_coef()
}
