% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ergm_fc.R
\name{get_ergm_fc}
\alias{get_ergm_fc}
\title{Convert ergm coefficients to fold changes}
\usage{
get_ergm_fc(ergm, trim = TRUE)
}
\arguments{
\item{ergm}{an ergm fit}

\item{trim}{whether to remove coefficients where the coefficient is -Inf}
}
\value{
tibble of fold changes
}
\description{
The fold changes are all based on comparison to reference level
}
\examples{
ergm <- get_ergms(example_network, preds = "site", types = "nodemix")
get_ergm_fc(ergm[[1]])
}
