% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ergms.R
\name{get_ergms}
\alias{get_ergms}
\title{get_ergms}
\usage{
get_ergms(network, preds = NULL)
}
\arguments{
\item{network}{igraph object}

\item{preds}{vector of predictors}
}
\value{
list of ergm models
}
\description{
Given node attributes and a network, fits all possible ergms
}
\examples{
get_ergms(example_network, c("site", "genetic_sex"))
}
