% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example.R
\name{get_example}
\alias{get_example}
\title{loads some example data}
\usage{
get_example(example = "example_network")
}
\arguments{
\item{example}{name of example data to upload}
}
\value{
example network
}
\description{
For use in the data wizard module
}
\examples{
get_example()
}
