% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lat_range.R
\name{get_lat_range}
\alias{get_lat_range}
\title{get lat or lon from network sf object}
\usage{
get_lat_range(network_sf, type = "lat")
}
\arguments{
\item{network_sf}{network sf object}

\item{type}{whether lat or lon}
}
\value{
2 vector of lat or lon
}
\description{
get lat or lon from network sf object
}
\examples{
get_lat_range(example_sf, "lon")
get_lat_range(example_sf, "lat")
}
