% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_measures.R
\name{get_network_measures}
\alias{get_network_measures}
\title{get_network_measures}
\usage{
get_network_measures(g, measure = "degree", var = "")
}
\arguments{
\item{g}{igraph object}

\item{measure}{centrality measure to use}

\item{var}{vertex variables to stratify on}
}
\value{
summary table
}
\description{
get_network_measures
}
\examples{
get_network_measures(example_network, var = "site")
}
