% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_summary.R
\name{get_network_summary}
\alias{get_network_summary}
\title{Get network summary}
\usage{
get_network_summary(g)
}
\arguments{
\item{g}{ibd network}
}
\value{
tibble with summary info on network
}
\description{
Get network summary
}
\examples{
get_network_summary(example_network)
}
