% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_info.R
\name{get_node_info}
\alias{get_node_info}
\title{Get network node information}
\usage{
get_node_info(g)
}
\arguments{
\item{g}{ibd network object}
}
\value{
tibble of node information
}
\description{
Get network node information
}
\examples{
get_node_info(example_network)
}
