% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ringbauer_measures.R
\name{get_ringbauer_measures}
\alias{get_ringbauer_measures}
\title{Get ringbauer measures}
\usage{
get_ringbauer_measures(g, grp)
}
\arguments{
\item{g}{igraph object}

\item{grp}{a name of a categorical node variable}
}
\value{
tibble with density measures
}
\description{
This takes a network and a categorical node variable and calculate the intra- and
inter- level density.
}
\examples{
get_ringbauer_measures(example_network, "site")
}
