% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_font.R
\name{get_text_font}
\alias{get_text_font}
\title{Get text font}
\usage{
get_text_font(x, type = "fill")
}
\arguments{
\item{x}{numeric vector to convert to colour}

\item{type}{whether to return the colour, or black / white for text}
}
\value{
vector of colours
}
\description{
Used to get contrasting text to a fill.
}
\examples{
get_text_font(1:10)
get_text_font(1:10, "text")
}
