% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_grp_duplicates.R
\name{label_grp_duplicates}
\alias{label_grp_duplicates}
\title{label_grp_duplicates}
\usage{
label_grp_duplicates(df, grp1, grp2)
}
\arguments{
\item{df}{data frame with two group columns}

\item{grp1}{name of first group col}

\item{grp2}{name of second group col}
}
\value{
df with extra col called .grp_label that has can be
used to remove duplicated rows
}
\description{
label_grp_duplicates
}
\examples{
df <- tidyr::expand_grid(
  grp1 = LETTERS[1:3],
  grp2 = LETTERS[1:3]
) |>
dplyr::mutate(
  value = 1:9
)
label_grp_duplicates(df, grp1, grp2)
}
