% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dist_network.R
\name{load_dist_network}
\alias{load_dist_network}
\title{load distance network}
\usage{
load_dist_network(node_file, edge_file)
}
\arguments{
\item{node_file}{file with meta data}

\item{edge_file}{file with edge list}
}
\value{
igraph object
}
\description{
load distance network
}
