% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ergm_bic.R
\name{plot_ergm_bic}
\alias{plot_ergm_bic}
\title{Plots BIC}
\usage{
plot_ergm_bic(
  ergms,
  text_size = 10,
  text_angle = 90,
  abbr = FALSE,
  measure = "BIC",
  top_5 = FALSE
)
}
\arguments{
\item{ergms}{a list of ergms}

\item{text_size}{size of x axis text}

\item{text_angle}{angle of x-axis text}

\item{abbr}{boolean to clean up names}

\item{measure}{decide if AIC or BIC.}

\item{top_5}{boolean to show just top 5 models}
}
\value{
plot
}
\description{
Final version is based on BR's version
}
\examples{
ergms <- example_network |>
  get_ergms(
    preds = c("site", "genetic_sex"),
    types = c("nodematch", "nodemix"))
plot_ergm_bic(ergms) |> print()
}
