% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ergm.R
\name{plot_ergm_coef}
\alias{plot_ergm_coef}
\title{Plot ergm coefficients}
\usage{
plot_ergm_coef(ergms, type = "theta", trim = TRUE)
}
\arguments{
\item{ergms}{list of ergm fits}

\item{type}{either "theta" or "phi" - fold changes}

\item{trim}{remove -Inf coefficients}
}
\value{
plot of coefficients
}
\description{
Plot ergm coefficients
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex")
)
ergms |> plot_ergm_coef()
}
