% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggnet.R
\name{plot_ggnet}
\alias{plot_ggnet}
\title{plot ggnet}
\usage{
plot_ggnet(
  ggnet_obj,
  fill_col = "",
  shape_col = "",
  node_size = 4,
  text_size = 4,
  labels = FALSE,
  label_col = "",
  label_inc = "",
  label_exc = "",
  connected = TRUE
)
}
\arguments{
\item{ggnet_obj}{ggnetwork dataframe}

\item{fill_col}{vertex attribute for node fill}

\item{shape_col}{vertex attribute for node shape}

\item{node_size}{node size}

\item{text_size}{label size}

\item{labels}{add labels}

\item{label_col}{vertex attribute to use for labels}

\item{label_inc}{regular expression to include labels}

\item{label_exc}{regular expression to exclude labels}

\item{connected}{boolean to include isolated nodes}
}
\value{
ggnetwork plot
}
\description{
plot ggnet
}
\examples{
example_ggnet <- ggnetwork::ggnetwork(example_network)
plot_ggnet(example_ggnet)
}
