% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-homophily.R
\name{plot_homophily}
\alias{plot_homophily}
\title{Plot homophily}
\usage{
plot_homophily(RM, p = NULL)
}
\arguments{
\item{RM}{Ringbauer tibble}

\item{p}{nework density}
}
\value{
Plot of homophily density
}
\description{
Plot homophily
}
\examples{
get_ringbauer_measures(example_network, "site") |> plot_homophily()
}
