% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_leaflet.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{plot leaflet}
\usage{
plot_leaflet(g, lat, lon, col = "site", label = "none", tile = "Default")
}
\arguments{
\item{g}{igraph object}

\item{lat}{vertex attribute that gives latitude}

\item{lon}{vertex attribute that gives longitude}

\item{col}{vertex attribute to give marker colour}

\item{label}{vertex attribute to give marker label}

\item{tile}{background tile}
}
\value{
leaflet plot
}
\description{
plot leaflet
}
\examples{
plot_leaflet(example_network, "lat", "long", col = "site")
}
