% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ringbauer.R
\name{plot_ringbauer}
\alias{plot_ringbauer}
\title{Plot Ringbauer}
\usage{
plot_ringbauer(RM, label_margin = 2, label_size = 3)
}
\arguments{
\item{RM}{a Ringbauer matrix object}

\item{label_margin}{the margin around the heatmap - adjust to get enough space for labels}

\item{label_size}{adjust label and axis text}
}
\value{
Ringbauer plot
}
\description{
Takes the Ringbauer matrix and plots it. This was based on the Ringbauer paper and also the heatmap.2 function from gplots
}
\examples{
get_ringbauer_measures(example_network, "site") |>
convert_ringbauer_measures() |>
plot_ringbauer(label_margin = 10, label_size = 3)
}
