% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_ergm_coef.R
\name{plot_single_ergm_coef}
\alias{plot_single_ergm_coef}
\title{Plot single ergm coefficients}
\usage{
plot_single_ergm_coef(ergm, type = "theta", trim = TRUE)
}
\arguments{
\item{ergm}{ergm model}

\item{type}{either "theta" or "phi" - fold changes}

\item{trim}{remove -Inf coefficients}
}
\value{
plot of coefficients
}
\description{
Plot single ergm coefficients
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex")
)
ergms[[1]] |> plot_single_ergm_coef()
}
