% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static_map.R
\name{plot_static_map}
\alias{plot_static_map}
\title{plot static map}
\usage{
plot_static_map(
  network_sf,
  zoom = NULL,
  maptype = "stamen_terrain",
  key = NULL,
  fill_col = "",
  shape_col = "",
  edge_col = "",
  lon_range = NULL,
  lat_range = NULL,
  theme = "minimal",
  pt_size = 3
)
}
\arguments{
\item{network_sf}{a network sf object}

\item{zoom}{level of tile resolution}

\item{maptype}{type of tile}

\item{key}{API key for stadia}

\item{fill_col}{node attribute for fill}

\item{shape_col}{node attribute for shape}

\item{edge_col}{edge attribute for edge width and colour}

\item{lon_range}{lon range to zoom}

\item{lat_range}{lat range to zoom}

\item{theme}{type of ggplot}

\item{pt_size}{node size}
}
\value{
plot of network with background tiles
}
\description{
plot static map
}
