% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_redundant_models.R
\name{remove_redundant_models}
\alias{remove_redundant_models}
\title{remove redundant models}
\usage{
remove_redundant_models(models, vars)
}
\arguments{
\item{models}{list of models}

\item{vars}{list of variables to filter on}
}
\value{
list so each var appears only once
}
\description{
remove redundant models
}
\examples{
M1 <- "network ~ edges + nodematch('A') + nodemix('A')"
M2 <- "network ~ edges + nodematch('A')"
remove_redundant_models(list(M1, M2), "A")
}
