% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_grp_duplicates.R
\name{sort_label}
\alias{sort_label}
\title{sort labels}
\usage{
sort_label(label1, label2)
}
\arguments{
\item{label1}{first label}

\item{label2}{second label}
}
\value{
label of form first-second
}
\description{
Takes two labels and sorts them so for example
A and B becomes A-B, and B and A becomes A-B.
This means that you can find duplicate edges in
an undirected network as they will have the same
label
}
\examples{
sort_label("A", "B")
sort_label("B", "A")
}
