% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spibder_app.R
\name{spibder_app}
\alias{spibder_app}
\title{spIBDerverse shiny app}
\usage{
spibder_app(input_network = NULL)
}
\arguments{
\item{input_network}{an igraph network object for the IBD data}
}
\value{
nothing
}
\description{
spIBDerverse shiny app
}
