% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ergm_coef.R
\name{tab_ergm_coef}
\alias{tab_ergm_coef}
\title{Tab ergm coefficients}
\usage{
tab_ergm_coef(ergms, models = NULL)
}
\arguments{
\item{ergms}{list of ergm fits}

\item{models}{models to show}
}
\value{
plot of coefficients
}
\description{
Tab ergm coefficients
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex"),
  types = c("nodematch", "nodemix")
)
ergms |> plot_ergm_coef()
}
