% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_alpha.R
\name{transform_alpha}
\alias{transform_alpha}
\title{Transform vector to (a,b)}
\usage{
transform_alpha(x, a = 0, b = 1)
}
\arguments{
\item{x}{vector to transform}

\item{a}{lower bound of transform}

\item{b}{upper bound of transform}
}
\value{
transformed vector
}
\description{
Transform vector to (a,b)
}
\examples{
transform_alpha(runif(100))
}
