% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sf.R
\name{convert_sf}
\alias{convert_sf}
\title{Convert to SF}
\usage{
convert_sf(g, lat, lon, jitter = 0)
}
\arguments{
\item{g}{igraph network}

\item{lat}{attribute that gives latitude}

\item{lon}{attribute that gives longitude}

\item{jitter}{amount of jitter to add}
}
\value{
list of edges_sf and nodes_sf
}
\description{
Converts an igraph object into two sf objects
}
\examples{
convert_sf(example_network,"lat","long")
}
