% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ibd_network.R
\name{create_ibd_network}
\alias{create_ibd_network}
\title{create_ibd_network}
\usage{
create_ibd_network(
  ibd_file,
  meta_file = NULL,
  ibd_co,
  frac_co = 0.49,
  only_connected = TRUE
)
}
\arguments{
\item{ibd_file}{TSV file with IBD data}

\item{meta_file}{TSV with node metadata}

\item{ibd_co}{vector of cutoffs for defining an edge}

\item{frac_co}{cutoff for quality control}

\item{only_connected}{boolean to remove unconnected nodes}
}
\value{
IBD network
}
\description{
create_ibd_network
}
