% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_network.R
\docType{data}
\name{example_network}
\alias{example_network}
\title{example network}
\format{
An igraph object
\describe{
\item{genetic}{Genetic sex for each sample with levels F or M}
\item{site}{Site for each sample with levels A, B, or C}
\item{name}{Name for each node}
\item{degree}{Degree for each node}
\item{lat}{Latitude for each node}
\item{long}{Longitude for each node}
\item{wij}{Weight for each edge}
}
}
\source{
Simulation via ergm package
}
\usage{
example_network
}
\description{
An igraph example network
}
\keyword{datasets}
