% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_network.R
\name{filter_network}
\alias{filter_network}
\title{Filter network}
\usage{
filter_network(
  g,
  node_inc = "",
  node_exc = "",
  node_column = "none",
  node_cutoff = NULL,
  edge_column = "none",
  edge_cutoff = NULL
)
}
\arguments{
\item{g}{ibd network}

\item{node_inc}{regular expression for inclusion}

\item{node_exc}{regular expression for exclusion}

\item{node_column}{node attribute to filter network on}

\item{node_cutoff}{cutoffs for node filter}

\item{edge_column}{edge attributes to filter network on}

\item{edge_cutoff}{cutoffs for edge filter}
}
\value{
filtered IBD network
}
\description{
Includes or excludes nodes based on regular expression
}
\examples{
filter_network(example_network, node_exc = "1")
}
