% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_attributes.R
\name{get_node_attributes}
\alias{get_node_attributes}
\title{get node attributes}
\usage{
get_node_attributes(g, type = "all")
}
\arguments{
\item{g}{network object}

\item{type}{type to return}
}
\value{
list of attributes names
}
\description{
Remove centrality measures and can filter on cat or num
}
\examples{
get_node_attributes(example_network)
}
