% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_centrality.R
\name{plot_centrality}
\alias{plot_centrality}
\title{plot centrality measures}
\usage{
plot_centrality(g, measure = "degree", facets = NULL)
}
\arguments{
\item{g}{igraph obj}

\item{measure}{centrality measure to plot}

\item{facets}{attributes to facet plot on}
}
\value{
histogram or barchart
}
\description{
plot centrality measures
}
\examples{
plot_centrality(example_network)
}
