% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ergm.R, R/plot_ergm_bic.R
\name{plot_ergm_bic}
\alias{plot_ergm_bic}
\title{plot_ergm_bic}
\usage{
plot_ergm_bic(ergms, text_size = 10, text_angle = 90, abbr = FALSE)

plot_ergm_bic(ergms, text_size = 10, text_angle = 90, abbr = FALSE)
}
\arguments{
\item{ergms}{a list of ergms}

\item{text_size}{size of x axis text}

\item{text_angle}{angle of x-axis text}

\item{abbr}{boolean to clean up names}
}
\value{
ggplot of BIC for each ergm

plot
}
\description{
For each ergm model plots the BIC

Final version is based on BR's version
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex")
)
ergms |> plot_ergm_bic()
ergms <- example_network |>
  get_ergms(c("site", "genetic_sex"))
plot_ergm_bic(ergms) |> print()
}
