% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ergm.R
\name{plot_ergm_bic}
\alias{plot_ergm_bic}
\title{plot_ergm_bic}
\usage{
plot_ergm_bic(ergms)
}
\arguments{
\item{ergms}{list of ergms}
}
\value{
ggplot of BIC for each ergm
}
\description{
For each ergm model plots the BIC
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex")
)
ergms |> plot_ergm_bic()
}
