% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ergm.R, R/plot_ergm_bic.R
\name{plot_ergm_bic}
\alias{plot_ergm_bic}
\title{plot_ergm_bic}
\usage{
plot_ergm_bic(ergms, type = 1)

plot_ergm_bic(ergms, type = 1)
}
\arguments{
\item{ergms}{list of ergms fitted to IBD net}

\item{type}{which plot to show}
}
\value{
ggplot of BIC for each ergm

plot of fits
}
\description{
For each ergm model plots the BIC
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex")
)
ergms |> plot_ergm_bic()
example_network |>
get_ergms(c("site", "genetic_sex")) |>
plot_ergm_bic()
}
