% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ergm.R, R/plot_ergm_coef.R
\name{plot_ergm_coef}
\alias{plot_ergm_coef}
\title{Plot ergm coefficients}
\usage{
plot_ergm_coef(ergms, type = "theta", trim = TRUE, models = NULL)

plot_ergm_coef(ergms, type = "theta", trim = TRUE, models = NULL)
}
\arguments{
\item{ergms}{list of ergm fits}

\item{type}{either "theta" or "phi" - fold changes}

\item{trim}{remove -Inf coefficients}

\item{models}{models to show}
}
\value{
plot of coefficients

plot of coefficients
}
\description{
Plot ergm coefficients

Plot ergm coefficients
}
\examples{
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex"),
  types = c("nodematch", "nodemix")
)
ergms |> plot_ergm_coef()
ergms <- get_ergms(
  example_network,
  preds = c("site", "genetic_sex"),
  types = c("nodematch", "nodemix")
)
ergms |> plot_ergm_coef()
}
