% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_homophily.R
\name{plot_homophily}
\alias{plot_homophily}
\title{Plot homophily}
\usage{
plot_homophily(RM, show_sign = FALSE, filter_sign = FALSE)
}
\arguments{
\item{RM}{Ringbauer tibble}

\item{show_sign}{boolean on whether to colour by significance}

\item{filter_sign}{boolean to show only significant edges}
}
\value{
Plot of homophily density
}
\description{
Plot homophily
}
\examples{
get_ringbauer_measures(example_network, "site") |> plot_homophily()
}
