% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_homophily.R
\name{plot_homophily}
\alias{plot_homophily}
\title{Plot homophily}
\usage{
plot_homophily(RM, g = NULL)
}
\arguments{
\item{RM}{Ringbauer tibble}

\item{g}{IBD network}
}
\value{
Plot of homophily density
}
\description{
Plot homophily
}
\examples{
get_ringbauer_measures(example_network, "site") |> plot_homophily()
}
