% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{plot network}
\usage{
plot_network(
  g,
  seed = 2026,
  connected = "Show",
  edge = "none",
  edge_legend = TRUE,
  edge_trans = "identity",
  label = "none",
  label_size = 4,
  label_inc = "",
  label_exc = "",
  fill = "none",
  shape = "none",
  node_size = 5,
  node_centrality = "none",
  pal = "ravenclaw"
)
}
\arguments{
\item{g}{network}

\item{seed}{seed to set node locations}

\item{connected}{choice for how to deal with isolated nodes with choices
Hide, Show, Grey out}

\item{edge}{edge attribute for line colour}

\item{edge_legend}{boolean to control edge legend}

\item{edge_trans}{transformation for edge mapping}

\item{label}{vertex attribute to use for labels}

\item{label_size}{label size}

\item{label_inc}{regular expression to include labels}

\item{label_exc}{regular expression to exclude labels}

\item{fill}{vertex attribute for node fill}

\item{shape}{vertex attribute for node shape}

\item{node_size}{node size}

\item{node_centrality}{vertext attribute for node alpha}

\item{pal}{colour palette}
}
\value{
network plot
}
\description{
plot network
}
\examples{
plot_network(example_network)
}
