% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{plot ggnet}
\usage{
plot_network(
  g,
  fill_col = "none",
  shape_col = "none",
  node_alpha_col = "none",
  edge_col = "none",
  edge_trans = "identity",
  edge_legend = TRUE,
  node_size = 4,
  text_size = 4,
  text_col = "black",
  labels = FALSE,
  label_col = "",
  label_inc = "",
  label_exc = "",
  connected = "Show"
)
}
\arguments{
\item{g}{network}

\item{fill_col}{vertex attribute for node fill}

\item{shape_col}{vertex attribute for node shape}

\item{node_alpha_col}{vertext attribute for node alpha}

\item{edge_col}{edge attribute for line colour}

\item{edge_trans}{transformation for edge mapping}

\item{edge_legend}{boolean to control edge legend}

\item{node_size}{node size}

\item{text_size}{label size}

\item{text_col}{text colour}

\item{labels}{add labels}

\item{label_col}{vertex attribute to use for labels}

\item{label_inc}{regular expression to include labels}

\item{label_exc}{regular expression to exclude labels}

\item{connected}{choice for how to deal with isolated nodes with choices
Hide, Show, Grey out}
}
\value{
network plot
}
\description{
plot ggnet
}
\examples{
plot_network(example_network)
}
