% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_staticmap.R
\name{plot_staticmap}
\alias{plot_staticmap}
\title{plot static-map}
\usage{
plot_staticmap(
  sf,
  key = NULL,
  zoom = 5,
  maptype = "stamen_terrain",
  lon_range = NULL,
  lat_range = NULL,
  fill = "none",
  shape = "none",
  node_size = 4,
  node_centrality = "none",
  connected = "Show",
  edge = "none",
  edge_legend = TRUE,
  edge_trans = "identity",
  label = "",
  label_inc = "",
  label_exc = "",
  label_size = 3,
  theme = "minimal"
)
}
\arguments{
\item{sf}{network sf pobject}

\item{key}{stadia API key}

\item{zoom}{stadia tile zoom}

\item{maptype}{stadia map tile}

\item{lon_range}{range of lon to zoom to}

\item{lat_range}{range of lat to zoom to}

\item{fill}{vertex attribute for node fill}

\item{shape}{vertex attribute for node shape}

\item{node_size}{node size}

\item{node_centrality}{vertext attribute for node alpha}

\item{connected}{choice for how to deal with isolated nodes with choices
Hide, Show, Grey out}

\item{edge}{edge attribute for line colour}

\item{edge_legend}{boolean to control edge legend}

\item{edge_trans}{transformation for edge mapping}

\item{label}{vertex attribute to use for labels}

\item{label_inc}{regular expression to include labels}

\item{label_exc}{regular expression to exclude labels}

\item{label_size}{label size}

\item{theme}{type of plot theme}
}
\value{
network plot
}
\description{
plot static-map
}
\examples{
plot_network(example_network)
}
